Make 51job detail parsing testable without source eval
The 51job detail parser and its supporting helpers now live in importable
TypeScript modules under apps/browser-extension/src/lib, while the legacy
content script is rebuilt from src/content.ts with esbuild.

This removes the brittle brace-matching eval test harness, adds richer
multi-entry fixtures plus unit/integration coverage, and treats content.js
as a generated artifact instead of hand-maintained source.

Constraint: Keep the browser extension runtime behavior intact while refactoring only the 51job parser slice
Constraint: content.js must still ship for the MV3 extension zip even after becoming generated output
Rejected: Harden the eval-based harness | still couples tests to monolithic source layout and string parsing
Rejected: Modularize the entire content script | broader risk than needed for this parser-focused change
Confidence: high
Scope-risk: moderate
Reversibility: clean
Directive: Build apps/browser-extension/content.js from src/content.ts before packaging or manual extension loading
Tested: cd apps/browser-extension && node esbuild.config.mjs
Tested: npm test
Tested: make check
Tested: ./scripts/build-extension-zip.sh
Tested: Live 51job detail extraction via window.__TR_RESUME_DATA__.extract() on April 2, 2026
Not-tested: Full 51job search-page auto-sync end-to-end after the new bundling flow
