+++ Makefile
@@ -441,7 +441,11 @@ build-static-fresh:
 	./scripts/build-static.sh --fresh
 
 # Build browser extension zip for web UI download
+build-extension:
+	cd apps/browser-extension && node esbuild.config.mjs
+
 build-extension-zip:
+	$(MAKE) build-extension
 	./scripts/build-extension-zip.sh
 
 # Serve static site locally (port 8000)
