+++ apps/browser-extension/package.json
@@ -9,12 +9,15 @@
     "cmux:setup-profile": "./scripts/cmux-setup-profile.sh",
     "open-search": "./scripts/open-search.sh",
     "check-profile-seed": "./scripts/check-profile-seed.sh",
+    "build": "node esbuild.config.mjs",
+    "build:watch": "node esbuild.config.mjs --watch",
     "lint": "eslint .",
     "typecheck": "tsc --noEmit"
   },
   "devDependencies": {
     "@eslint/js": "^9.13.0",
     "chrome-types": "^0.1.327",
+    "esbuild": "^0.27.0",
     "eslint": "^10.1.0",
     "globals": "^17.4.0",
     "typescript": "~5.6.2"
