+++ apps/browser-extension/popup-auto-sync-summary.test.mjs
@@ -1,5 +1,3 @@
-/* global console */
-
 import { readFileSync } from 'node:fs';
 import assert from 'node:assert/strict';
 import path from 'node:path';
