+++ apps/browser-extension/src/content.ts
@@ -1,3 +1,27 @@
+// @ts-nocheck
+import {
+  filterResumesByAgeRange,
+  getAgeRangeFromUrl,
+  normalizeOptionalPositiveInt,
+} from "./lib/job51-age-filter";
+import {
+  EHIRE_51JOB_HOST,
+  EHIRE_51JOB_PROFILE_URL_PREFIX,
+  buildJob51DetailResumeFromPayload,
+  normalizeJob51MultilineText,
+  normalizeJob51Text,
+} from "./lib/job51-detail-parser";
+import {
+  resolveJob51AutoSyncDetailWaitMode,
+  resolveJob51CollectionLimits,
+  resolveJob51DetailFetchDelayMs,
+} from "./lib/job51-collection-config";
+import {
+  buildWorkHistoryRawParts,
+  normalizeResumeMultilineText,
+  normalizeResumeText,
+} from "./lib/resume-text-utils";
+
 /**
  * 智通直聘 Resume Collector - Content Script
  * Extracts resume data from hr.job5156.com/search page
@@ -52,11 +76,8 @@ const AUTO_LOCATION_PARAM = "location";
 const AUTO_KEYWORD_MODE_PARAM = "tr_kw_mode";
 const SAMPLE_NAME_PARAM = "tr_sample_name";
 const JOB5156_HOST = "hr.job5156.com";
-const EHIRE_51JOB_HOST = "ehire.51job.com";
 const SEEK_HOST_SUFFIX = ".employer.seek.com";
 const JOB5156_PROFILE_URL_PREFIX = `https://${JOB5156_HOST}/resume/view/`;
-const EHIRE_51JOB_PROFILE_URL_PREFIX =
-  "https://ehire.51job.com/Revision/talent/resume/detail?contentType=&resumeId=";
 const SOURCE_KEYS = {
   JOB5156: "job5156",
   JOB51: "51job",
@@ -66,8 +87,6 @@ const SOURCE_KEYS = {
 const SEEK_PROFILE_TYPE = "seek";
 const KEYWORD_MODE_CONCAT = "concat";
 const KEYWORD_MODE_SPACED = "spaced";
-const JOB51_SAFE_LIMIT = 50;
-const JOB51_SAFE_MAX_PAGES = 1;
 const JOB51_PAGE_COOLDOWN_MS = 8000;
 const JOB51_RATE_LIMIT_ERROR_MESSAGE =
   "51job 已触发访问频率限制，请60分钟后再试";
@@ -84,8 +103,6 @@ const JOB5156_DETAIL_FETCH_TIMEOUT_MS = 5000;
 const JOB5156_DETAIL_FETCH_CONCURRENCY = 5;
 const JOB51_DETAIL_FETCH_TIMEOUT_MS = 8000;
 const JOB51_DETAIL_FETCH_CONCURRENCY = 2;
-const JOB51_DETAIL_FETCH_DELAY_MS = 5000;
-const JOB51_DETAIL_FETCH_UNSAFE_DELAY_MS = 1000;
 const DEFAULT_SEEK_PAGE_SIZE = 20;
 const LATEST_AUTO_SYNC_SUMMARIES_STORAGE_KEY = "latestAutoSyncSummaries";
 const DEFAULT_COLLECTION_GUARDS = {
@@ -152,63 +169,6 @@ function normalizeKeywordMode(mode) {
 function normalizeCollectionLimit(value) {
   const parsed = Number.parseInt(String(value ?? ""), 10);
   return Number.isFinite(parsed) && parsed > 0 ? parsed : 0;
-}
-
-function normalizeOptionalPositiveInt(value) {
-  const parsed = Number.parseInt(String(value ?? "").trim(), 10);
-  if (!Number.isFinite(parsed) || parsed <= 0) return null;
-  return parsed;
-}
-
-function parseAgeNumber(value) {
-  if (typeof value === "number" && Number.isFinite(value) && value > 0) {
-    return Math.trunc(value);
-  }
-  if (typeof value !== "string") return null;
-  const trimmed = value.trim();
-  if (!trimmed) return null;
-
-  const withSuffix = trimmed.match(/(\d+)\s*岁/u);
-  if (withSuffix && withSuffix[1]) {
-    const parsed = Number.parseInt(withSuffix[1], 10);
-    return Number.isFinite(parsed) && parsed > 0 ? parsed : null;
-  }
-
-  const plainNumber = trimmed.match(/^(\d{1,3})$/u);
-  if (plainNumber && plainNumber[1]) {
-    const parsed = Number.parseInt(plainNumber[1], 10);
-    return Number.isFinite(parsed) && parsed > 0 ? parsed : null;
-  }
-
-  return null;
-}
-
-function getAgeRangeFromUrl() {
-  const params = new URLSearchParams(window.location.search || "");
-  const minAge = normalizeOptionalPositiveInt(params.get(AUTO_MIN_AGE_PARAM));
-  const maxAge = normalizeOptionalPositiveInt(params.get(AUTO_MAX_AGE_PARAM));
-  const enabled = minAge !== null || maxAge !== null;
-  return {
-    enabled,
-    minAge: minAge !== null ? minAge : undefined,
-    maxAge: maxAge !== null ? maxAge : undefined,
-  };
-}
-
-function filterResumesByAgeRange(resumes) {
-  const range = getAgeRangeFromUrl();
-  if (!range.enabled) return resumes;
-
-  const minAge = range.minAge;
-  const maxAge = range.maxAge;
-
-  return resumes.filter((resume) => {
-    const age = parseAgeNumber(resume?.age);
-    if (age === null) return false;
-    if (typeof minAge === "number" && age < minAge) return false;
-    if (typeof maxAge === "number" && age > maxAge) return false;
-    return true;
-  });
 }
 
 const PROVINCE_TOKENS = new Set([
@@ -305,7 +265,11 @@ async function getCollectionLimits() {
           : normalizeCollectionLimit(items?.maxPages);
         if (getCurrentSourceKey() === SOURCE_KEYS.JOB51) {
           resolve(
-            resolveJob51CollectionLimits(resolvedLimit, resolvedMaxPages),
+            resolveJob51CollectionLimits(
+              resolvedLimit,
+              resolvedMaxPages,
+              window.location.search,
+            ),
           );
           return;
         }
@@ -324,6 +288,7 @@ async function getCollectionLimits() {
           resolveJob51CollectionLimits(
             hasLimitParam ? paramLimit : 0,
             hasMaxPagesParam ? paramMaxPages : 0,
+            window.location.search,
           ),
         );
         return;
@@ -334,43 +299,6 @@ async function getCollectionLimits() {
       });
     }
   });
-}
-
-function hasJob51UnsafeLimitsOverride() {
-  const params = new URLSearchParams(window.location.search || "");
-  return params.get("tr_unsafe_limits") === "1";
-}
-
-function resolveJob51CollectionLimits(limit, maxPages) {
-  if (hasJob51UnsafeLimitsOverride()) {
-    return {
-      limit: limit > 0 ? limit : JOB51_SAFE_LIMIT,
-      maxPages: maxPages > 0 ? maxPages : JOB51_SAFE_MAX_PAGES,
-    };
-  }
-  return {
-    limit: limit > 0 ? Math.min(limit, JOB51_SAFE_LIMIT) : JOB51_SAFE_LIMIT,
-    maxPages:
-      maxPages > 0
-        ? Math.min(maxPages, JOB51_SAFE_MAX_PAGES)
-        : JOB51_SAFE_MAX_PAGES,
-  };
-}
-
-function resolveJob51DetailFetchDelayMs() {
-  return hasJob51UnsafeLimitsOverride()
-    ? JOB51_DETAIL_FETCH_UNSAFE_DELAY_MS
-    : JOB51_DETAIL_FETCH_DELAY_MS;
-}
-
-function resolveJob51AutoSyncDetailWaitMode() {
-  const params = new URLSearchParams(window.location.search || "");
-  const mode = normalizeResumeText(params.get("tr_job51_detail_wait") || "")
-    .toLowerCase();
-  if (mode === "page1" || mode === "all") {
-    return mode;
-  }
-  return "background";
 }
 
 function buildExportFilename() {
@@ -1192,53 +1120,6 @@ function buildSeekProfileEducationItem(item) {
     qualification: qualification || undefined,
     endDate: endDate || undefined,
   };
-}
-
-function normalizeResumeText(value) {
-  return typeof value === "string"
-    ? value.replace(/[\u3000\s]+/g, " ").trim()
-    : "";
-}
-
-function stripHtmlTags(value) {
-  return typeof value === "string" ? value.replace(/<[^>]*>/g, " ") : "";
-}
-
-function normalizeJob51Text(value) {
-  return normalizeResumeText(stripHtmlTags(value));
-}
-
-function normalizeJob51MultilineText(value) {
-  return normalizeResumeMultilineText(stripHtmlTags(value));
-}
-
-function isLikelyJob51LocationPlaceholderCompanyName(value) {
-  const text = normalizeJob51Text(value);
-  if (!text) return false;
-  if (
-    /(公司|集团|科技|机械|工业|实业|设备|自动化|贸易|精密|制造|电子|机电|工具|刀具|技术|股份|责任|有限|厂|大学|学院|学校|中心|医院|门诊|商贸|材料|模具|液压|传感)/u
-      .test(text)
-  ) {
-    return false;
-  }
-  if (!/^[\u4e00-\u9fa5]{2,4}$/u.test(text)) {
-    return false;
-  }
-  return true;
-}
-
-function normalizeResumeMultilineText(value) {
-  if (typeof value !== "string") return "";
-  return value
-    .replace(/\r\n?/g, "\n")
-    .split("\n")
-    .map((line) => line.replace(/[\u3000\t ]+/g, " ").trim())
-    .filter(Boolean)
-    .join("\n");
-}
-
-function buildWorkHistoryRawParts(parts) {
-  return parts.filter(Boolean).join(" · ");
 }
 
 function parseJob5156BasicInfoItems(items, locationOverride = "") {
@@ -2409,7 +2290,7 @@ async function enrich51JobSearchResumesWithDetail(resumes, options = {}) {
     typeof options.interBatchDelayMs === "number" &&
     Number.isFinite(options.interBatchDelayMs)
       ? options.interBatchDelayMs
-      : resolveJob51DetailFetchDelayMs();
+      : resolveJob51DetailFetchDelayMs(window.location.search);
   const collectionGuards = await loadCollectionGuards();
   const guardFields = parseGuardFieldNames(
     collectionGuards?.[SOURCE_KEYS.JOB51],
@@ -3573,633 +3454,6 @@ function extract51JobResumes() {
   });
 }
 
-function getJob51DetailRoot(payload) {
-  if (!payload) return null;
-  if (Array.isArray(payload)) {
-    for (const entry of payload) {
-      const root = getJob51DetailRoot(entry);
-      if (root) return root;
-    }
-    return null;
-  }
-  if (typeof payload !== "object") return null;
-  const record = payload;
-  const candidates = [
-    record.data,
-    record.result,
-    record.resume,
-    record.detail,
-    record.resumeInfo,
-    record.resumeDetail,
-    record.resumeViewVo,
-    record.resume_view_vo,
-    record.resumeVo,
-    record.cnVo,
-    record.content,
-  ];
-  for (const candidate of candidates) {
-    const root = getJob51DetailRoot(candidate);
-    if (root) return root;
-  }
-  return record;
-}
-
-function readJob51Text(...values) {
-  for (const value of values) {
-    if (typeof value === "string") {
-      const text = normalizeJob51Text(value);
-      if (text) return text;
-    } else if (typeof value === "number" && Number.isFinite(value)) {
-      return String(value);
-    }
-  }
-  return "";
-}
-
-function readJob51MultilineText(...values) {
-  for (const value of values) {
-    if (typeof value === "string") {
-      const text = normalizeJob51MultilineText(value);
-      if (text) return text;
-      continue;
-    }
-    if (Array.isArray(value)) {
-      const text = value
-        .map((entry) =>
-          typeof entry === "string"
-            ? normalizeJob51MultilineText(entry)
-            : entry && typeof entry === "object"
-              ? readJob51MultilineText(
-                  entry.text,
-                  entry.value,
-                  entry.content,
-                  entry.description,
-                  entry.desc,
-                  entry.detail,
-                  entry.duty,
-                  entry.responsibility,
-                )
-              : "",
-        )
-        .filter(Boolean)
-        .join("\n");
-      if (text) return text;
-    }
-  }
-  return "";
-}
-
-function normalizeJob51DateLike(value) {
-  const text = readJob51Text(value);
-  if (!text) return "";
-  if (["至今", "目前", "今"].includes(text)) return "至今";
-  return text
-    .replace(/[./年]/g, "-")
-    .replace(/月/g, "")
-    .replace(/-{2,}/g, "-")
-    .replace(/^-|-$/g, "");
-}
-
-function readJob51Array(record, keys) {
-  if (!record || typeof record !== "object") return [];
-  for (const key of keys) {
-    const value = record[key];
-    if (Array.isArray(value)) return value;
-  }
-  return [];
-}
-
-function buildJob51ExperienceEntry(item, kind) {
-  if (!item || typeof item !== "object") return null;
-
-  const isProject = kind === "project";
-  const rawCompanyName = readJob51Text(
-    item.company_name,
-    item.companyName,
-    item.compname,
-    item.comp_name,
-    item.com_name,
-    item.comName,
-    item.company,
-    isProject ? item.project_name : undefined,
-    isProject ? item.projectName : undefined,
-    isProject ? item.project : undefined,
-  );
-  const companyName = isLikelyJob51LocationPlaceholderCompanyName(rawCompanyName)
-    ? ""
-    : rawCompanyName;
-  const jobTitle = readJob51Text(
-    item.work_func_value,
-    item.workfunc,
-    item.workfunc_str,
-    item.work_func_str,
-    item.job_name,
-    item.jobName,
-    item.position,
-    item.jobTitle,
-    isProject ? item.project_role : undefined,
-    isProject ? item.role : undefined,
-    isProject ? item.responsibility : undefined,
-  );
-  const startDate = normalizeJob51DateLike(
-    item.start_time ??
-      item.startDate ??
-      item.begin ??
-      item.start_date ??
-      item.fromDate ??
-      item.time_begin ??
-      item.timefrom,
-  );
-  const endDate = normalizeJob51DateLike(
-    item.end_time ??
-      item.endDate ??
-      item.end ??
-      item.end_date ??
-      item.toDate ??
-      item.time_end ??
-      item.timeto,
-  );
-  const durationLabel = readJob51Text(
-    item.working_years,
-    item.worktime,
-    item.duration,
-    item.timeDiff,
-    item.time_diff,
-    item.period,
-    item.durationLabel,
-  );
-  const description = Array.from(
-    new Set(
-      [
-        item.workdescribe,
-        item.work_describe,
-        item.workDescription,
-        item.work_content,
-        item.workContent,
-        item.work_detail,
-        item.workDetail,
-        item.workDuty,
-        item.work_duty,
-        item.duty,
-        item.duties,
-        item.responsibility,
-        item.responsibilities,
-        item.responsibility_text,
-        item.responsibilityText,
-        item.responsibility_list,
-        item.responsibilityList,
-        item.duty_list,
-        item.dutyList,
-        item.description,
-        item.desc,
-        item.detail,
-        item.content,
-        item.achievement,
-        item.achievements,
-        isProject ? item.project_desc : undefined,
-        isProject ? item.projectDescribe : undefined,
-      ]
-        .map((value) => readJob51MultilineText(value))
-        .filter(Boolean),
-    ),
-  ).join("\n");
-  const metaParts = [
-    item.industry_tag,
-    item.workindustry,
-    item.company_size_value,
-    item.companysize,
-    item.work_type_value,
-    item.companytype,
-    item.section,
-    item.department,
-    item.project_name,
-  ]
-    .map((value) => readJob51Text(value))
-    .filter(Boolean);
-  const raw = buildWorkHistoryRawParts([
-    [startDate, endDate].filter(Boolean).join("~"),
-    durationLabel ? `(${durationLabel})` : "",
-    companyName,
-    jobTitle,
-    ...metaParts,
-    description,
-  ]);
-
-  if (!raw && !description && !companyName && !jobTitle) return null;
-
-  return {
-    raw:
-      raw ||
-      description ||
-      buildWorkHistoryRawParts([companyName, jobTitle, startDate, endDate]),
-    companyName: companyName || undefined,
-    jobTitle: jobTitle || undefined,
-    description: description || undefined,
-    startDate: startDate || undefined,
-    endDate: endDate || undefined,
-  };
-}
-
-function buildJob51EducationEntry(item) {
-  if (!item || typeof item !== "object") return null;
-
-  const institution = readJob51Text(
-    item.school_name,
-    item.schoolName,
-    item.schoolname,
-    item.school,
-    item.institution,
-    item.university,
-    item.college,
-  );
-  const qualification = readJob51Text(
-    item.degree_value,
-    item.degreename,
-    item.degree,
-    item.degreeStr,
-    item.qualification,
-    item.education,
-  );
-  const fieldOfStudy = readJob51Text(
-    item.major,
-    item.degreemajor,
-    item.major_name,
-    item.speciality,
-    item.field_of_study,
-    item.subject,
-  );
-  const description = readJob51MultilineText(
-    item.describe,
-    item.description,
-    item.detail,
-    item.content,
-  );
-  const startDate = normalizeJob51DateLike(
-    item.start_date ?? item.begin ?? item.startTime ?? item.enrollYear ?? item.timefrom,
-  );
-  const endDate = normalizeJob51DateLike(
-    item.end_date ?? item.end ?? item.endTime ?? item.graduationYear ?? item.timeto,
-  );
-
-  if (!institution && !qualification && !fieldOfStudy && !description) {
-    return null;
-  }
-
-  return {
-    institution: institution || undefined,
-    qualification: qualification || undefined,
-    fieldOfStudy: fieldOfStudy || undefined,
-    description: description || undefined,
-    startDate: startDate || undefined,
-    endDate: endDate || undefined,
-  };
-}
-
-function buildJob51SkillEntry(item) {
-  if (typeof item === "string") {
-    const text = normalizeJob51Text(item);
-    return text || null;
-  }
-  if (!item || typeof item !== "object") return null;
-
-  const name = readJob51Text(
-    item.skill_name,
-    item.name,
-    item.label,
-    item.skill,
-    item.tag,
-  );
-  if (!name) return null;
-
-  const level = readJob51Text(item.level, item.skill_level, item.proficiency);
-  const yearsOfExperience =
-    typeof item.yearsOfExperience === "number" &&
-    Number.isFinite(item.yearsOfExperience)
-      ? item.yearsOfExperience
-      : typeof item.years === "number" && Number.isFinite(item.years)
-        ? item.years
-        : typeof item.years === "string" && item.years.trim()
-          ? item.years.trim()
-          : undefined;
-
-  if (!level && yearsOfExperience === undefined) {
-    return name;
-  }
-
-  return {
-    name,
-    ...(level ? { level } : {}),
-    ...(yearsOfExperience === undefined ? {} : { yearsOfExperience }),
-  };
-}
-
-function buildJob51LicenceEntry(item) {
-  if (typeof item === "string") {
-    const text = normalizeJob51Text(item);
-    return text ? { name: text } : null;
-  }
-  if (!item || typeof item !== "object") return null;
-
-  const name = readJob51Text(
-    item.name,
-    item.cert_name,
-    item.certificate_name,
-    item.training_name,
-    item.course_name,
-    item.title,
-  );
-  if (!name) return null;
-
-  const authority = readJob51Text(
-    item.authority,
-    item.organization,
-    item.issuing_org,
-    item.issuingOrganisationName,
-    item.school,
-    item.company,
-  );
-  const issuedAt = normalizeJob51DateLike(
-    item.issuedAt ??
-      item.issue_date ??
-      item.issued_date ??
-      item.start_date ??
-      item.startTime,
-  );
-  const expiresAt = normalizeJob51DateLike(
-    item.expiresAt ??
-      item.expire_date ??
-      item.expired_date ??
-      item.end_date ??
-      item.endTime,
-  );
-
-  return {
-    name,
-    ...(authority ? { authority } : {}),
-    ...(issuedAt ? { issuedAt } : {}),
-    ...(expiresAt ? { expiresAt } : {}),
-  };
-}
-
-function buildJob51DetailResumeFromPayload(payload, options = {}) {
-  const root = getJob51DetailRoot(payload);
-  if (!root) return [];
-
-  const resumeId = readJob51Text(
-    options.resumeId,
-    root.resumeId,
-    root.resume_id,
-    root.userid,
-    root.userId,
-    root.user_id,
-    root.id,
-    root.base_info?.userid,
-    root.base_info?.resumeId,
-  );
-  const perUserId = readJob51Text(
-    root.accountid,
-    root.accountId,
-    root.account_id,
-    root.base_info?.accountid,
-    root.base_info?.accountId,
-    root.base_info?.account_id,
-    root.userid,
-    root.userId,
-    root.user_id,
-  );
-  const profileUrl = normalizeJob51Text(
-    options.profileUrl ||
-      (resumeId
-        ? `${EHIRE_51JOB_PROFILE_URL_PREFIX}${encodeURIComponent(resumeId)}`
-        : ""),
-  );
-  const baseInfo =
-    root.base_info && typeof root.base_info === "object" ? root.base_info : {};
-  const liveJobIntention =
-    Array.isArray(root.jobintention) && root.jobintention[0] &&
-    typeof root.jobintention[0] === "object"
-      ? root.jobintention[0]
-      : {};
-  const liveHighestDegree =
-    root.highestdegree && typeof root.highestdegree === "object"
-      ? root.highestdegree
-      : {};
-  const jobIntentionInfo =
-    root.job_intention && typeof root.job_intention === "object"
-      ? root.job_intention
-      : {};
-  const recentWorkInfo =
-    root.recent_work_info && typeof root.recent_work_info === "object"
-      ? root.recent_work_info
-      : {};
-  const workHistory = readJob51Array(root, [
-    "work",
-    "work_list",
-    "workInfoVoList",
-    "workInfoList",
-    "work_info",
-    "work_info_list",
-    "workHistory",
-    "work_history",
-    "work_experience",
-    "workExperience",
-    "workExperienceList",
-    "work_exp_list",
-  ])
-    .map((item) => buildJob51ExperienceEntry(item, "work"))
-    .filter(Boolean);
-  const projectExperience = readJob51Array(root, [
-    "project",
-    "project_list",
-    "projectInfoVoList",
-    "projectInfoList",
-    "projectExperience",
-    "project_experience",
-    "projectExperienceList",
-  ])
-    .map((item) => buildJob51ExperienceEntry(item, "project"))
-    .filter(Boolean);
-  const profileEducation = readJob51Array(root, [
-    "education",
-    "education_list",
-    "educationInfoVoList",
-    "profileEducation",
-    "educationHistory",
-  ])
-    .map((item) => buildJob51EducationEntry(item))
-    .filter(Boolean);
-  const skills = readJob51Array(root, [
-    "itskill",
-    "skill",
-    "skills",
-    "skill_list",
-    "label_sorted_skill_tag_list",
-    "label_list",
-  ])
-    .map((item) => buildJob51SkillEntry(item))
-    .filter(Boolean);
-  const licences = [
-    ...readJob51Array(root, [
-      "certification",
-      "certifications",
-      "certificate",
-      "certificates",
-    ]),
-    ...readJob51Array(root, ["train", "training", "train_list", "trainings"]),
-  ]
-    .map((item) => buildJob51LicenceEntry(item))
-    .filter(Boolean);
-
-  const name = readJob51Text(
-    root.name,
-    root.userName,
-    root.user_name,
-    root.username,
-    root.resume_name,
-    root.realName,
-    root.candidateName,
-    root.fullName,
-    baseInfo.resume_name,
-    baseInfo.name,
-    baseInfo.userName,
-  );
-  const age = readJob51Text(root.age, root.realAge, root.displayage, baseInfo.age);
-  const experience = readJob51Text(
-    baseInfo.work_year_value,
-    baseInfo.workYear,
-    baseInfo.workYears,
-    root.work_year_value,
-    root.workyear,
-    root.workYear,
-    root.workYears,
-    root.experienceYears,
-    root.experience,
-    recentWorkInfo?.working_years,
-  );
-  const education = readJob51Text(
-    baseInfo.top_degree_value,
-    root.top_degree_value,
-    liveHighestDegree.degree,
-    root.education,
-    root.degree,
-    root.degreeValue,
-  );
-  const location = readJob51Text(
-    jobIntentionInfo.expect_job_area_value,
-    Array.isArray(liveJobIntention.expectarea)
-      ? liveJobIntention.expectarea
-          .map((item) =>
-            item && typeof item === "object"
-              ? readJob51Text(item.provincecity, item.county)
-              : "",
-          )
-          .filter(Boolean)
-          .join(",")
-      : undefined,
-    root.jobIntention?.expect_job_area_value,
-    baseInfo.area_value,
-    root.area,
-    root.areaprovincecity,
-    root.location,
-    root.workCity,
-    root.city,
-    root.workLocation,
-  );
-  const jobIntention = readJob51Text(
-    jobIntentionInfo.expect_work_function_value,
-    jobIntentionInfo.expected_work_function_value,
-    liveJobIntention.expectfuncname,
-    liveJobIntention.expectposition,
-    root.jobIntention?.expect_work_function_value,
-    root.jobIntention?.expected_work_function_value,
-    recentWorkInfo.recent_position,
-    root.recentwork?.workname,
-    root.jobIntention,
-    root.job_name,
-    root.jobname,
-    root.desiredJob,
-    root.expectedPosition,
-    root.targetJob,
-    root.searchJob,
-  );
-  const expectedSalary = readJob51Text(
-    jobIntentionInfo.new_expect_salary,
-    jobIntentionInfo.expect_salary,
-    liveJobIntention.newdisplayexpectsalary,
-    liveJobIntention.displayexpectsalary,
-    root.expectedSalary,
-    root.desiredSalary,
-    root.expectSalary,
-    root.salaryStr,
-    root.salary,
-  );
-  const activityStatus = readJob51Text(
-    root.active_type,
-    root.activityStatus,
-    root.activetimelabel,
-    root.activetime,
-    root.lastLoginTime,
-    root.last_login_time,
-    baseInfo.active_type,
-    baseInfo.jobStateStr,
-  );
-  const selfIntro = readJob51MultilineText(
-    root.selfintro,
-    root.selfIntro,
-    root.resume_slicing,
-    root.advantage,
-    root.profile,
-    root.summary,
-    root.resumeSummary,
-    root.highlight,
-    root.professionSkill,
-    jobIntentionInfo.professionSkill,
-  );
-  const normalizedAge = age ? (age.includes("岁") ? age : `${age}岁`) : "";
-  const externalId = resumeId || perUserId;
-  const pageIndex = 1;
-  const source = EHIRE_51JOB_HOST;
-
-  if (
-    !resumeId &&
-    !name &&
-    !jobIntention &&
-    !selfIntro &&
-    workHistory.length === 0
-  ) {
-    return [];
-  }
-
-  return [
-    {
-      name,
-      age: normalizedAge,
-      experience,
-      education,
-      location,
-      jobIntention,
-      expectedSalary,
-      activityStatus,
-      selfIntro,
-      resumeId: resumeId || undefined,
-      perUserId: perUserId || undefined,
-      externalId: externalId || undefined,
-      profileUrl: profileUrl || undefined,
-      source,
-      workHistory,
-      projectExperience:
-        projectExperience.length > 0 ? projectExperience : undefined,
-      profileEducation:
-        profileEducation.length > 0 ? profileEducation : undefined,
-      skills: skills.length > 0 ? skills : undefined,
-      licences: licences.length > 0 ? licences : undefined,
-      pageIndex,
-      rawData: root,
-      extractedAt: new Date().toISOString(),
-    },
-  ];
-}
-
 function extractJob51DetailResume() {
   if (!isJob51DetailPage() || !isJob51DetailReady()) return [];
   return filterResumesByAgeRange(
@@ -4208,15 +3462,28 @@ function extractJob51DetailResume() {
         new URL(window.location.href).searchParams.get("resumeId") || undefined,
       profileUrl: window.location.href,
     }),
+    window.location.search,
+    AUTO_MIN_AGE_PARAM,
+    AUTO_MAX_AGE_PARAM,
   );
 }
 
 function extractResumes() {
   if (isJob51DetailPage()) {
-    return filterResumesByAgeRange(extractJob51DetailResume());
+    return filterResumesByAgeRange(
+      extractJob51DetailResume(),
+      window.location.search,
+      AUTO_MIN_AGE_PARAM,
+      AUTO_MAX_AGE_PARAM,
+    );
   }
   if (getCurrentSourceKey() === SOURCE_KEYS.JOB51) {
-    return filterResumesByAgeRange(extract51JobResumes());
+    return filterResumesByAgeRange(
+      extract51JobResumes(),
+      window.location.search,
+      AUTO_MIN_AGE_PARAM,
+      AUTO_MAX_AGE_PARAM,
+    );
   }
   if (getCurrentSourceKey() === SOURCE_KEYS.SEEK) {
     if (isSeekProfileMode()) {
@@ -4231,7 +3498,12 @@ function extractResumes() {
   }
 
   if (isJob5156DetailPage()) {
-    return filterResumesByAgeRange(extractJob5156DetailResume());
+    return filterResumesByAgeRange(
+      extractJob5156DetailResume(),
+      window.location.search,
+      AUTO_MIN_AGE_PARAM,
+      AUTO_MAX_AGE_PARAM,
+    );
   }
 
   const cards = document.querySelectorAll(SELECTORS.resumeCard);
@@ -4252,7 +3524,12 @@ function extractResumes() {
     }
   });
 
-  return filterResumesByAgeRange(resumes);
+  return filterResumesByAgeRange(
+    resumes,
+    window.location.search,
+    AUTO_MIN_AGE_PARAM,
+    AUTO_MAX_AGE_PARAM,
+  );
 }
 
 /**
@@ -5748,7 +5025,11 @@ function resolveAgeFilterActions(selectBox) {
 
 async function autoApplyAgeFilterFromUrl() {
   const sourceKey = getCurrentSourceKey();
-  const range = getAgeRangeFromUrl();
+  const range = getAgeRangeFromUrl(
+    window.location.search,
+    AUTO_MIN_AGE_PARAM,
+    AUTO_MAX_AGE_PARAM,
+  );
   if (!range.enabled) {
     setAutoAgeAttributes("skipped");
     return;
@@ -6244,7 +5525,9 @@ function queueJob51DetailBackfill(resumes, context = {}) {
     runId !== null && runId !== job51DetailBackfillRunId;
 
   const task = async () => {
-    const detailFetchDelayMs = resolveJob51DetailFetchDelayMs();
+    const detailFetchDelayMs = resolveJob51DetailFetchDelayMs(
+      window.location.search,
+    );
 
     if (isCancelled()) {
       console.log("51job detail backfill skipped", {
@@ -6515,7 +5798,11 @@ async function runAutoSyncIfEnabled() {
         resumes = await enrichSeekRecommendedResumesWithDetail(resumes);
       }
       if (resumes.length <= 0) {
-        const ageRange = getAgeRangeFromUrl();
+        const ageRange = getAgeRangeFromUrl(
+          window.location.search,
+          AUTO_MIN_AGE_PARAM,
+          AUTO_MAX_AGE_PARAM,
+        );
         const ageHint = ageRange.enabled
           ? ` · 年龄: ${typeof ageRange.minAge === "number" ? ageRange.minAge : "—"}-${typeof ageRange.maxAge === "number" ? ageRange.maxAge : "—"}`
           : "";
@@ -6624,7 +5911,9 @@ async function runAutoSyncIfEnabled() {
           currentPage,
           totalPages: Math.max(totalPages, currentPage),
         });
-        const waitMode = resolveJob51AutoSyncDetailWaitMode();
+        const waitMode = resolveJob51AutoSyncDetailWaitMode(
+          window.location.search,
+        );
         const shouldWaitForDetails =
           waitMode === "all" || (waitMode === "page1" && currentPage === 1);
         if (shouldWaitForDetails) {
@@ -7132,7 +6421,11 @@ async function collectSnapshotPayload(options = {}) {
 function getExternalAccessorStatus() {
   const version = getExtensionVersion();
   const pagination = getPaginationInfo();
-  const ageRange = getAgeRangeFromUrl();
+  const ageRange = getAgeRangeFromUrl(
+    window.location.search,
+    AUTO_MIN_AGE_PARAM,
+    AUTO_MAX_AGE_PARAM,
+  );
   const sourceKey = getCurrentSourceKey();
   const apiSnapshotCount = getApiSnapshotCount();
   const cardCount =
